requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
asyncpg==0.29.0
httpx==0.27.0
apscheduler==3.10.4
anthropic==0.21.3
python-dotenv==1.0.1
sqlalchemy==2.0.28
